feat: Rewrite VectorMath.cs with System.Numerics.Vector<float> for .NET Framework 4.8.1 compatibility
- Replace System.Runtime.Intrinsics with System.Numerics.Vector<T> API
- Maintain SIMD acceleration (2-4x throughput) compatible with Framework
- Remove AVX2-specific code paths (DotProductAvx2, NormAvx2, etc)
- Add Vector.IsHardwareAccelerated runtime detection
- Fix sp_Hypothesize.sql to use Atom table instead of non-existent CodeAtom
- Reorder Build-WithSigning.ps1: build DACPAC before signing (Step 3→4)
- Fix build-dacpac.ps1 to copy DACPAC from bin/Output to bin/Release
- Fix deploy-dacpac.ps1: add DropConstraintsNotInSource=True
- Fix Sign-CLRAssemblies.ps1: handle empty DLL directories gracefully
- Clean up nullable reference warnings (CS8601, CS8604, CS8618)
- Remove InferenceTracking table references from sp_Analyze.sql

All 127 CLR functions deploy successfully. VectorMath verified with strong-name signed dependencies (MathNet.Numerics, System.Numerics.Vectors).

+++ src/Hartonomous.Api/DTOs/Provenance/SearchStatistics.cs
@@ -5,5 +5,5 @@ public class SearchStatistics
     public long CandidatesEvaluated { get; set; }
     public double IndexHitRate { get; set; }
     public double CacheHitRate { get; set; }
-    public int AverageEmbeddingTime { get; set; }
+    public long AverageEmbeddingTime { get; set; }
 }
