+++ src/Hartonomous.Api/DTOs/Research/ResearchQueryResponse.cs
@@ -6,7 +6,7 @@ namespace Hartonomous.Api.DTOs.Research;
 public class ResearchQueryResponse
 {
     public string Query { get; set; } = string.Empty;
-    public int ExecutionTimeMs { get; set; }
+    public long ExecutionTimeMs { get; set; }
     public List<ResearchResult> Results { get; set; } = new();
     public QueryAggregations Aggregations { get; set; } = new();
     public SpatialBounds SpatialCoverage { get; set; } = new();
